requirements.txt
absl-py
aspose-slides
PyMuPDF